+++ tests/ci/services/smoke/smoke-profiles.util.spec.ts
@@ -12,7 +12,6 @@ describe("smoke-profiles.util", () => {
 				"use-memo.md",
 				"react-labs-view-transitions-activity-and-more.md",
 				"lazy.md",
-				"invalid-hook-call-warning.md",
 			]);
 		});